python/requirements.txt
fastapi[all]>=0.75
uvicorn[standard]>=0.15
pytest==8.3.4
Pillow>=8.0.0  # Pillowの追加